Dataviews: Add some client side data handling utils
