+++ packages/dataviews/src/index.js
@@ -1,2 +1,3 @@
 export { default as DataViews } from './dataviews';
+export { sortByTextFields, getPaginationResults } from './utils';
 export { VIEW_LAYOUTS } from './constants';
